More uses of qualified refs in parallel.cpp
The parallel pass was doing a poor job of identifying and considering
references for arg bundles. After a lot of thrash, it largely came down
to improving the decision-making for when autoCopies were inserted.
There's much more to be done, but I think this is a step in the right
direction and makes much more sense.

These changes caused a number of issues downstream, which required
addressing previously uncommon AST patterns.

This change also adds the FLAG_COFORALL_INDEX_VAR flag to a yielded
index for coforall loops. Before the information could have been lost
during lowerIterators.

Another side effect is a potential reduction in autoCopies for some
tests.