Move c_void_ptr comparisons from CPtr to ChapelBase
Since c_void_ptr is a primitive type, it makes sense that operations on
it should be in ChapelBase. CPtr.chpl is really for the c_ptr type.

The == and != operations join the = operation for c_void_ptr in
ChapelBase.

Also improved the testing for these functions to check both comparisons
with nil and also comparisons with other c_void_ptrs.