Index temps should be local to loop bodies in promotion wrappers
While building up a promotion wrapper we would create VarSymbols for
indices and insert a DefExpr at the top of the wrapper function, even
though the indices were only used in the soon-to-be-created loop bodies.
This resulted in cases where we would have dead code and could pass
uninitialized references around.

To fix this, we instead insert UnresolvedSymExprs and rely on
destructureIndices to create the VarSymbol and DefExpr within the loop's
body. After we have created the wrapper, we need to resolve the
UnresolvedSymExpr in the call to the function that we are promoting.

This allows me to also enable an optimization/bug fix in parallel.cpp
that avoids taking references to syncs on stacks that are about to go
away, and has the side-effect of reducing comm counts.