Resolve memory leak with standalone iteration
This change resolves the second memory leak with the test

 test/associative/bharshbarg/domains/domainSetOps.chpl

When iterExpr referred to a domain argument, the PRIM_MOVE to create the
variable chpl__iterSA was performing a deep copy when no copy should be
performed. Adding FLAG_MAYBE_REF to the temp chpl__iterSA allows the rest
of the compiler to infer that chpl__iterSA should be a reference in this
case.