Adjust jglewis dataflow cholesky tests for range task intent
After the commit named "Adjust jglewis cholesky tests to avoid array use
after free", the begin statements in these dataflow tests captured ranges
instead of temporary array slices. It currently happens to be the case
that ranges are passed to task functions with 'const in' intent by default.
However, this behavior is not specified and in the future we might want
it to be 'const ref' by default. So, adjust the code I added to these
tests to explicitly use the 'in' task intent for these ranges.