Allow paratest to work on whiteboxes and real Cray hardware
Paratest forwards the chplenv, explicitly setting the values on the clients.
However, on Cray's and whiteboxes we didn't allow CHPL_TARGET_ARCH to be set
manually, it had to be set through the cpu targeting module. This updates
chpl_arch.py to allow CHPL_TARGET_ARCH to be manually set to the value of the
cpu targeting module.

We also have to use chpl_launchcmd.py on some Cray's, but previously we had no
way of setting it for paratest. Now it can be set through CHPL_TEST_LAUNCHCMD

Lastly, nightly didn't respect CHPL_NIGHTLY_TEST_DIRS for paratest, so update
it to create a DIRFILE, like we do for -examples.