Limit whitebox paratesting to only run 1 executable at a time
I was setting CHPL_TEST_LIMIT_RUNNING_EXECUTABLES inside a function that got
called inside of a subshell, so it wasn't actually getting set for nightly
testing.

We now set it at the top-level of common-localnode-paratest.bash so it gets set
when that script is sourced. Clients will now have to manually unset
CHPL_TEST_LIMIT_RUNNING_EXECUTABLES if they don't want to limit the number of
executables running concurrently.