Make it easier to paratest a subset of the test suite
Currently it's a huge pain to paratest a subset of the test suite. paratest
supports a -DIRFILE option, that lets you specify a file that contains a list
of directories to test. If you just want to test release/examples and LAPACK
(what nightly XC testing uses) you have to do something like:

    find release/examples/ modules/standard/LAPACK -type d > DIRFILE
    ../util/test/paratest.server -dirfile DIRFILE

this adds a -dirs option to paratest that lets you specify a space separated
list of directories to recursively search for directories to test. Instead of
the above, you can now just do:

    ../util/test/paratest.server -dirs 'release/examples modules/standard/LAPACK'

or

    ../util/test/paratest.server -dirs release/examples -dirs modules/standard/LAPACK