Update attribution of these bigint versions
When pushing these to the release, I was focused on the code more
than the attribution of them.  Here I've updated the bigint versions
to be based on the GMP Chapel version and the LeDrug version and
given credit to Thomas (who did the original BigInt ports), Mike
(who updated Thomas's versions to bigints), and myself (who merged
them with the release GMP versions).  Sorry for not catching this
comment before the 1.14 release went out, guys.