Fix tuple issue, work around memory leak in digits.chpl
test/io/ferguson/digits.chpl was leaking memory. This memory leak turns out to
be due to an existing problem (that occurs on master) with eliminated casts -
but the code pattern occurs here because of tuple improvements. The test

  test/types/string/ferguson/string-cast-leak.chpl

shows the issue. Here I adjust IO.chpl to work around the problem in the case
of readf which is used by digits.chpl. That fixes the memory leak.

While debugging that issue, I discovered another problem with tuples and the
var (a,b) = someExpression; syntax. The tests in

  test/types/tuple/ferguson/tuple-string-bug*.chpl

explore some variations of the related code and I made a workaround change to
build.cpp that I believe covers most of the cases without adding a leak. Before
the change to build.cpp, some of these new tests would core dump.