Transposed c-ray and parallelized rendering step
This commit:

a) transposes the pixel array and
b) parallelizes the render step

The hope is that both changes will accelerate valgrind and baseline
testing to solve the timeout issues that have been occurring since
I checked in the test.  If not, I will reduce the test size.

The transpose of the pixel array reflects the fact that I have a
hard time remembering which order image dimensions are given in
and decided to transpose them for the Chapel array in order to
make the first and second dimension match the way we normally
draw Chapel arrays.  This also results in a more natural iteration
order when writing the image out to a file and therefore should
result in a better memory order in rendering (not that memory is
likely to be a bottleneck here).

The parallelization of the rendering step is trivial -- changed
a 'for' to a 'forall'.

While here, I noticed some arguments that weren't used due to
arrays/domains carrying their bounds with them.