Eliminate zippered iteration in pidigits
We were computing the index for the current digit in both the gen_digits iter
and at the call site. This just updates the iter to yield a tuple of the digit
and it's index.

This has the benefit of allowing the iterator to be inlined, although I didn't
see any performance benefit from this