Replace the pidigits loop with a more readable one
Convert a for-loop with doubly nested while-loops instead a single while-loop
with conditionals.

I think this is a lot easier to read, and it's the result of a couple of
iterations with Brad. I originally had a version that looked more like the
reference ledrug version (used continues instead of conditionals), but this is
cleaner.