Remove reindex call added by the compiler
calls to reindex and chpl__getDomainFromArrayType were added by the
compiler in a case like this:

class C {
  var A: [0..4] int;
}

var GA: [1..5] int = [i in 1..5] i;

var c2 = new C(A=>GA);  // <- note A=>GA here

See also test_array_alias_field.chpl

This pattern is used in dsiSlice in distributions, including BlockDist.

The call to chpl__getDomainFromArrayType had significant performance
overhead for miniMD, but it is unnecessary since we don't actually need
to reindex these array alias fields in practice. Now the type declaration
on A in the example above is still required for the aliasing behavior,
but will be otherwise ignored.