Optimize locking in slicing, aliasing, reindexing
Motivated by slicing in particular.

The implementation of slicing in ChapelArray creates a new domain.
Previously it would make a locking call to newDom.add_arr and also
call chpl_incRefCountsForDomainsInArrayEltTypes on the new array.

Now it makes a call to add_arr with locking=false, since locking
is not necessary to update a domain that we have just created
and not yet shared.

Additionally, it no longer calls chpl_incRefCountsForDomainsInArrayEltTypes,
since a slice/alias/reindex has shorter lifetime than a the original
array (or else program error). Therefore, any inner domain that
needs to be preserved will be handled by the original array. In order
to support that mechanism, no longer decrements this count in
_delete_arr if it's an alias/slice.