Remove autoCopy call from handleArrayTypeCase
This autoCopy call appears to no longer be necessary and was the reason
for test failures with these tests:

   classes/bradc/arrayInClass/genericArrayInClass-otharrs.chpl
   classes/bradc/arrayInClass/genericClassArray-named.chpl

since autoCopy now creates a complete copy of an array
(whereas before this arrays effort it would do reference counting).

Additionally, since runtime type records are never destroyed in any way,
the result of the autoCopy call was leaked.

By removing the autoCopy, this change causes runtime types to capture
shallow copies of domain/distribution wrapper records.  That amounts to
capturing a domain or distribution by reference, but might be confusing.
I made an initial attempt to adjust the runtime type to store a reference
field but that seemed to create other issues during function resolution.