Avoid adding duplicate filenames
Presumably this could've always caused issues, but it particularly
does now that we process require statements in function resolution.
Specifically, in the case of a require statement in a generic,
we would process the same statement multiple times and add
multiple #includes for a .h file which would break if the .h file
wasn't designed to support that.  Similarly, it seems it could
either break, or at least waste effort for, duplicate .c/.h files.