Improve opaque domains to work-around strange problem
Before this commit, opaque domains would use a class instance
to be the opaque index. This commit changes it to a record
that has an integer field that is generated by a local atomic.

This works around a problem with the test
  arrays/opaque/bradc/opaqueIterate.chpl
which for some reason is triggering a bug in lowerIterators.
The test works if iterator inlining is disabled and also works
if the for loop deleting opaque indexes is deleted from the
opaque domain destructor.

I think it's smarter to use the record-of-integers strategy
anyway. The atomic fields are per-locale, so generating an opaque
index does not require communication.