Split fasta and fasta-redux cases into two graphs
This splits fasta and fasta-redux into two separate graphs.  We had
not been distinguishing well between them before (maybe the CLBG
hadn't been either?), and had been betting recently that fasta-redux
would persist, but it hasn't.  The one version that straddles this
divide is fasta-blc.chpl which originally was a fastaredux code but
recently was converted into a fasta code (to be annotated in a future
commit)