Improve release fasta: use int probabilities/randoms
As I improve my study version of fasta, my plan is to propagate
changes into the release version as they prove profitable.  Here,
I'm incorporating the change (used by the #1 entry) of having
the RNG generate ints rather than floats, and to pre-scan the
cumulated probabilities to convert them from floating point values
to int values in order to avoid floating point math in the inner
loop.

As part of this change, I did some refactorings, some necessary,
some for aesthetics:

* moved the sumProbs() function into the randomMake() routine since
  it's related to that routine.

* removed the nested addLine() procedure which seemed pretty
  pointless.

* moved the RNG parameters to config params, in part because one
  of them has to be used in multiple scopes now, in part because
  these characterize the benchmarks.

These changes correspond to the ones I made in my study version in PR
parallelism adds a lot of code overhead and didn't result in a clear
win as far as I could tell, so I'm going to save that one for last in
the release version after making other improvements that seem to help
my study version.