When do record copies occur document
I've made many versions of this document. This version represents the
most recent version but edited for the current implementation information
on the arrays branch.

It does not necessarily read well. I'm intending it to be here for
historical reference and as a starting point to updating the language
specification.