Split array-reduce-intents to working test and future
The part that does not work is the user-defined PlusReduceOp when the array is
distributed with Block. In that case, there is an apparent function resolution
issue when resolving PlusReduceOp.identity.  the visible functions do not
include PlusReduceOp.identity and it appears to be a problem where the visible
functions are not correctly based upon the point of instantiation.

Note that the arrays effort adds calls to array assignment/initCopy that are
resolved always. In this way, the DefaultRectangular these() array iterator was
already resolved by the time that Block tried to use it in this test. I think
that this earlier version is being found but is not properly respecting the new
visible functions.