Update CHIP status expectations in 1.rst, update statuses
I've found it hard to say when a CHIP is "Accepted" but often
can say more in the status field. This commit adjusts the expectations
for the status field described in 1.rst to be something I think is
more achievable to track and then updates the statuses for various
CHIPs to indicate what is going on with them.