Fix for failure to recognize field initialization in block statements
The issue was that our check about the left hand side of a call to "=" only
applied to statements that were at the top level scope of the initializer
functions.  If they occur in a block statement, we would skip over the block
statement entirely instead of diving into the statements within its scope.  This
change alters the iteration style of phase1Analysis to also traverse into the
nested blocks (as long as they aren't loop statements, which I will handle
next).  I will need to also apply something similar to this change when
searching for "this.init()" and "super.init()" calls, but that's a bit of a
separate effort.