Add an SPMD-style variant to LCALS
This new variant runs each kernel by:
  * starting maxTaskPar tasks in a coforall loop
  * using the RangeChunk package to chunk up the loop space for the tasks
  * Adding an explicit barrier at the end of the num_samples loop

This means it only has to create and clean up maxTaskPar tasks instead of
num_samples*maxTaskPar tasks per kernel, where num_samples tends to be around
15,000.

Because it doesn't have a way to reduce parallelism for the SHORT and MEDIUM
loop sizes, they still take too long to run.  For now, the program will halt
early if attempting to run the SPMD variant with SHORT or MEDIUM sizes.

Because the SHORT and MEDIUM sizes are disabled, this entire variant is
disabled by default.  To turn it on, turn off SHORT and MEDIUM:

a.out --run_shortLoops=false --run_mediumLoops=false --run_variantRawSPMD=true

This variant gets many of the kernels much closer to matching the performance
of the parallel reference version.