Make learnChapelInYMinutes.prediff not care about stdout/stderr ordering.
The learnChapelInYMinutes test produces one line of stderr output during
its serial section.  The .prediff used to care about where that line was
with respect to all the stdout output from the serial section.  But the
ordering of stderr with respect to stdout depends on everything from
shells to launchers that aren't under Chapel's control.  So, allow that
one line of stderr to occur in any order with respect to the serial
output.  It actually could occur anywhere with respect to the parallel
output also, but we aren't having trouble with that yet, so don't worry
about it for now.

This change is being done as part of the GASNet 1.28.0 upgrade because
somewthing changed in the amudprun launcher's stdout/stderr buffering
in that release such that the test started failing.  With this commit it
works again.