Adjust test of ref-pair functions
I think the change in .bad file represents an improvement.
While there, I updated the .chpl file to no longer have comments
for the cases that don't match expectation. Then, I updated
the .good file to match my expectation. The difference here
is that I'm pretty sure that a function taking in a 'const ref'
argument should not trigger the 'setter' version.