Remove usesOuterVars from virtual method call resolution
I started looking at usesOuterVars because of an issue with
several tests in darwin + GASNet testing. E.g.

  release/examples/benchmarks/ssca2/SSCA2_main

fails when clang compiles the generated C code:

  tmp/ChapelArray.c:24886:544: fatal error: bracket nesting level exceeded

In ChapelArray.chpl, there is:

  proc _freePrivatizedClass(pid:int, original:object):void {
    ...

    delete original

    ...
  }

This is being code-generated into a function that is about 10,000 lines long
and consists of a whole lot of nested conditionals (implementing 'delete
original').

During resolution of dynamic-dispatch calls, insertDynamicDispatchCalls calls
usesOuterVars which is meant to check if an overridden method refers to an
outer variable. If it does refer to an outer variable, it emits a series of
conditionals instead of a virtual method table call. There is a problem with
usesOuterVars where it returns true in the case above (and when it presumably
should not). So, the above case creates a lot of conditionals.

It looks like this behavior started in the commit adding record-based strings
to Chapel: 21b548aeb3. Since usesOuterVars does not seem to have changed, it
seems likely that the new strings revealed a latent bug in usesOuterVars.

usesOuterVars was originally added in f54ca676 as part of first-class function
support. It enables two tests to pass:

  functions/jturner/lambda_capture_var
  trivial/jturner/lambda_parent_capture_issue

First-class-fn support needs to be revisited/rewritten and several language
design questions need answering. One language design question is how
first-class-functions interact with captured variables. The language does not
necessarily need to support captured variables in first-class functions at all.

Since the language design for first-class-functions is not complete, this patch
simply removes support for first-class-functions that capture variables and
makes the two tests into futures. This issue needs to be revisited along with a
clear language design.