Fix comm counts again
In PR #4792, I assumed that comm counts would be the same on different
machines. That is not currently the case. So, this PR updates the .good
files to match the output on the machines where nightly testing is run.

I suspect that the issue is the same one that Ben has identified as a
performance concern for STREAM - that tuples containing arrays are containing
references to arrays and so are doing a GET when the _array record is
dereferenced. Hopefully we can solve this by improving remote value forwarding.