Improve const-refs and RVF
RVF thinks it can forward an argument that is a const-ref, as long as no
code needs to know the original locale of that argument. There are a
number of things in our AST that can be inferred to be const-ref, but
are instead marked as a plain 'ref'.

I've added the mini-pass 'inferConstRefs' that will walk the AST and
attempt to determine if ArgSymbols or VarSymbols with QUAL_REF (or
INTENT_REF) can be changed to QUAL_CONST_REF or INTENT_CONST_REF.

I've also needed to add analysis to RVF itself to determine if a
const-ref argument may be used in a ".locale" query. If it is, then we
should not RVF that argument.

This work should resolve some of the increased comm-counts introduced
PR4745, which added qualified references to the compiler.