Move fences for on statement calls to parallel
The direct local on optimization was originally adding these fences
within the non-local branch of the if statement.  But, when that code was
refactored into parallel from createTaskFunctions, the fences were added
around the whole conditional.

This commit changes that behavior so that the fences are again added
inside theh non-local branch. To do so, it adds two new flags that
communicate from createTaskFunctions to passArgsToNestedFns in parallel
that a fences should be added around the call.