Fix bugs when reading null bytes into strings
Chapel strings store a length (unlike C strings) and so can store
strings that contain zero bytes. However, we had some bugs preventing
this from working when reading a particular number of bytes from a file
into a string.

The bugs had to do with the construction of the string from the
c_string_copy that the I/O system created. In particular,
we were calling

    str_out = new string(tx, needToCopy=false);

which causes the string's length to be initialized by finding the
zero byte (which is normally what you want for C strings). But the
I/O system read a particular number of bytes and returned the size
of the buffer. So, instead we now call

    str_out = new string(tx, length=readlength, needToCopy=false);

and have added this length= argument to that string constructor.