Fix incorrect handling of field initialization within a for loop (in basic case)
Prior to this change, we wouldn't traverse for loops looking for field
initialization.  Mostly importantly, this meant that we didn't generate an error
message for those initializations, as field initialization in loops is not
allowed.  Of secondary importance, we wouldn't even behave as the user expected
us to, because we would consider the fields initialization in the loop as
"omitted" and so insert initialization for them prior to the next field
initialized, thereby overwriting any initialization performed in the loop.  By
detecting and then erroring out, the user should see the correct behavior
(which is an error when this occurs)