Use export keyword with chpl_nodeFromLocaleID etc
The other routines in the LocaleModels.chpl file are marked with export
but these were not for some reason. I believe that was an oversight.

One of the effects of marking them with export is that the arguments are
not considered to be wide. (note that an alternative way to keep the
arguments from being widened is to use pragma "local args").

For some reason, PR #4794 caused testing failures in numa configuration.
These failures were C-level compilation failures because
chpl_nodeFromLocaleID was passed a local pointer actual when the type of
the formal was a wide pointer. I'm not sure why PR #4794 triggered these
errors, but the solution is pretty clear. The functions now 'export'd are
there to create or unpack locale IDs and are not meant to include
communication.