Make discovery of the next statement to check happen all in one place.
Prior to this, I had put getting the next statement back up the tree in one
place and going down into the tree in another.  That seemed unnecessarily
complex.  So I altered getNextStmt to check if the "next" statement to look at
was a block statement or loop and if so, dive into it (depending on if we
thought searching loops was valid at this point).  I had to add a fencepost
dive to the beginning of phase1Analysis, but other than that, all of this
checking lives in one place.

phase1Analysis traversal doesn't dive into loops on its own - instead, if a
loop is returned to it, we pass that along to verifyLoopIsClean, which behaves
differently when field initialization is found (and knows to dive into any loops
it finds).

I suppose I could move the fencepost portion within getNextStmt, too.  But I
think I won't for right now.