RVF record-wrapped fields within records
Now that record-wrapped types are always RVF'd, we can safely RVF
record-wrapped fields within records (provided the top-level record
itself is RVF'd).

This is accomplished by creating a wrapper/bundle record for the
on-statement, and bit-copying the record-wrapped fields into this
bundle. Within the on-statement, the bundle is unpacked and we overwrite
the original reference fields with references to the local RVF'd
record-wrapped data.

This work was motivated by increased comm-counts as a result of PR 4762,
which improved array memory management and disabled the unwrapping of
array records.