Update nextField instead of the argument curField in verifyLoopIsClean
Ugh.  Shouldn't have resulted in incorrect behavior, just the unnecessary
creation and return of a variable.  I was consistent enough in using the wrong
one that I didn't run into an infinite loop.

Also update curField with the returned value in phase1Analysis' call to it,
since we aren't updating it in the body of the callee any more.