Couple new tests
As happened with my field initialization tests, my test of super.init() call
detection in loops was missing the call because it also was within a conditional
statement.  Add a test that includes the call at the top scope of the loop body.

Also add a test where a loop was encountered after we'd already seen field
initialization for every field.  To ensure full detection, I made this loop
contain a duplicate field initialization.  Right now, there is a bug in my
iteration set up so I miss this case.