Fix missing error message with loops occurring after all fields have been seen
Prior to this change, I failed to dive into loops when we had encountered all
fields at least once.  I also did not adjust for finding the last field in the
loop but still having statements remaining within it.  This meant that there
were some error messages that were not getting printed out, and there was the
potential for segfaults when I tried to access information on the expected next
field if there were no more expected fields.  I solved this by adding a check
that the field was not NULL before accessing its members in the other checks,
and by adding my loop search call to the branch of phase1Analysis when no more
fields are expected.

With this change, and a slight modification to the .good file (forthcoming),
the test hasLoopBad3 now passes.