Edit expected output for hasLoopBad3 slightly
I need to reshuffle my error cases messages slightly, but I haven't thought of
a good way to do that yet.  In an ideal world, I would handle the error
message in there as well (and generate the loop message in addition to the
other messages when both apply), but that doesn't happen quite yet.