Fix for getVisibleFunctions issue
When implementForallIntents2() specializes an iterator or wrapper
for specific forall intents, the FnSymbol->instantiationPoint
of the specialization used to be inherited from the original.
This caused resolution errors when the forall loop was accessing
or referring to, esp. via reduce intents, something that was not
visible from the original context.

This change makes instantiationPoint point to the new context.

Background: instantiationPoint affects the chain of visibility
blocks - see getVisibilityBlock(), used in getVisibleFunctions().

The new test, written by Michael:

  test/parallel/forall/vass/visibility-of-record-for-reduce.chpl

illustrates that. The forall that implements the three "+ reduce"
expressions invokes the default rectangular array standalone
iterator. The first time that iterator is instantiated is from doB().
so its instantiationPoint points to doB's block. For the one in doC(),
it specializes the iterator instantiated for doB(). Before this change,
the instantiationPoint of the specialization pointed to doB as well,
so the bar record and its related functions were not visible.
With this change, instantiationPoint is adjusted to point into doC.