Add an option --[no-]print-callgraph to print the callgraph for programs
--print-callgraph causes a textual callgraph to be printed just after
resolution.  Recursive calls are marked "(Recursive) fn_name" and stop
the traversal down a given path.  Repeated calls to the same function
from a given function only print once, including calls to different
instantiations of a generic function.

This printout needs to be done after resolution, but before the cleanup
that happens at the end of resolution. This is because it needs all calls
to be resolved, but needs to happen before generic functions are removed
as unused.  This way all instantiations of a function can be treated as
a single function - the generic.