Simplify arg handling / --help
Turns out that some gasnet failures last night were because argv[0]
for a gasnet program is './foo_real' rather than './foo'.  This
suggests some future work:

  - either the '_real' executable should lie about the executable
    name, particularly if/when it can determine that the launcher
    was used.

  - or better, the launcher program should handle any argv arguments
    as it does configs.  Specifically, we wouldn't want to sit in
    a slurm/PBS queue only to finally run and find out that the user
    had thrown --help

Because of these issues, I decided to stop processing argv arguments
and just add a config named 'help' (throwing away the option to also
trigger on 'h') to control whether or not help is printed out.
Unfortunately, this didn't work because our built-in help flag takes
precedence over a user config named help.  Sigh...  So then I called
the config usage.  So some other future work might be:

  - to have a user 'help' config take precedence over the built-in
    one;

  - or perhaps ditto any other user config that conflicts with a
    built-in

Anyway, in spite of the frustrations, this results in a simpler
main() and gets rid of some arguable overengineering that I did
in my usage/help routine to try and stick to 80 columns of code
and to be flexible w.r.t. error vs. --help cases.

Removed a test case that tested for bad arguments since it no
longer does anything specific to this test.