Future against how/when argv[0] is printed out
Argv[0] currently prints out 'foo_real' when a launcher is used, which
seems inappropriate since that's not what the user is typing.
Moreover, it seems unfortunate that the argv processing isn't run in
the launcher to avoid sitting in a queue before doing it.  This
future captures that notion -- see the .future for more.