Three fixes to for_vector etc
I noticed three problems for_vector, for_set, and for_queue in stlUtil.h.

First, these macros declared what appears to be the loop variable outside
of the loop. That means that an example such as this:

      for_vector(Symbol, sym, symbolVector) {
        ...
      }
      for_vector(Symbol, sym, symbolVector) {
        ...
      }

does not compile, because now there are two 'sym' variables declared in
the same scope. (It could be fixed if the second loop changes the index
variable to sym2 for example).

Second, these macros evaluated the collection being iterated over more
than once. Coverity pointed this out (in an obtuse manner) for code in
loopInvariantCodeMotion, such as:

  for_set(Symbol, rhsAlias, aliases[rhs]) {
    ..
  }

Third, the macro for_queue is never used in the compiler. It is dead
code.

This PR addresses these 3 problems. It adjusts the macros for
for_vector and for_set to declare the iterator as a variable outside
of the loop but to do so with a unique-ified name (by appending the
line number to theh variable name). Additionally, it adjusts these
macros to only evaluate the collection being iterated over once. The
loop variable (e.g. sym or rhsAlias in the examples above) is declared
in the for loop since its name cannot be changed.

Lastly (and most trivially) this PR removes the for_queue macro
since:
 * any adjustment to it would not be tested, and
 * adding it back would be fairly trivial based upon for_vector/for_set