Fix remaining (?) c-ray test issues
Last night's remaining c-ray test failures seem to be due to a
fragility that I inadvertantly introduced, combined with the test
system considering tests in different orders on different systems
(presumably due to locale settings and the like).  This commit
should fix those cases.

Specifically, I had a directory-wide CATFILES option which made sense
when I was only testing correct executions, but once I started testing
error cases, if those ran first (as they did on linux32 but not
linux64), the file would not yet exist, causing an error message to be
printed (and conversely, unfortunately, when they ran second, the file
would just happen to still be hanging around from the correct cases so
the CATFILES file was being satisfied inappropriately).  This raises
questions for me about whether the test system should honor CATFILES
options for tests that fail in compilation, but the simpler fix was to
tighten up my testing.  To that end, here, I:

* split the remaining error case tests and usage test into their own
  .chpl files which 'use' c-ray.chpl to avoid code duplication.

* move the directory-wide CATFILES option into c-ray.catfiles so that
  it's only invoked for correct executions.