When printing call graphs, start at the compiler generated main, not user-main
The compiler generated main includes the call to the module init function,
which will be where the program really starts if there was no explicit main.
The generated main also includes the call to the user-main if there is one.

Don't print anything or increase the indent level for the generated main
or for the module init function, hiding their existence.