Added an ADT for forall intents replacing byrefVars and PRIM_FORALL_LOOP. Note: 'byrefVars' are used for task intents as before.
RATIONALE
=========

For a easy-to-deal-with and robust implementation of reductions and
especially partial reductions, we need to move away from the current
approach. It stems from lowering, in the parser, of forall loops and
iterators (somewhat). It causes heavy reliance on AST pattern-matching,
which leads to complex and brittle code and slow progress.

This PR is my first step towards "dedicated" representation of forall
loops in the compiler from the parser through resolve(). I expect that
the fields of the current ForallIntents class will become fields
of a new ForallLoop node. At which point ForallIntents will go away.

HIGHLIGHTS
==========

* Added ForallIntents class. It contains a list of variables
subject to forall intents, with the corresponding intent tag
and, for reduce spec, the reduce-intent expression - what
appears before 'reduce' in the Chapel source.

* Each BlockStmt that represents the body of a forall loop will point
to a ForallIntents via BlockStmt::forallIntents.  Prior to this PR,
this was done via BlockStmt::byrefVars pointing to a PRIM_FORALL_LOOP
call-expr containing an encoding of the vars and their intents.

* Changed parser and build.cpp to generate/manipulate ForallIntents things
directly instead of PRIM_ACTUALS_LIST call-exprs.

* The parser reflects, in a ForallIntents, only those variables that
the user listed in a with-clause explicitly. implementForallIntents1(),
currently run within resolve(), adds outer variables that are subject
to the default forall intents.

* Retired PRIM_FORALL_LOOP.

* BlockStmt::byrefVars is still used for task intents.

* tiMark* things are still used for the same purpose as before,
they are just introduced later in compliation instead of
coming from the parser. This is an intermediate situation.
TODO eliminate them entirely.

* Added enum TFITag for Task/Forall Intents.
Cf. enum IntentTag does not provide reduce intents.
It makes sense to me to keep the two sets of intents separate.
However obviously we can make it work if we merge them into one.

* Did not add new tests. We already have multiple tests
for forall/reduce intents.

DETAILS
=======

* In addition to the list of intent-ed variables, ForallIntents
also keeps three SymExprs. These should also go away eventually.
I had these three in PRIM_FORALL_LOOP call-exprs, so transferred them
to ForallIntents for now.

* Augmented AstVisitor hierarchy with visitForallIntents(ForallIntents*).

* Invariant: forallIntents is null for subclasses of BlockStmt.
Because those subclasses do not correspond to actual blocks.

* Invariant: forallIntents is null for BlockStmt after resolution.
Because during resolution all forall intents are converted away.

* Invariant: riSpecs[i] is null for non-reduce intent positions
i.e. when !isReduce(i). fiVars[i] are always non-null.

* BaseAST::remove() on an expression directly under ForallIntents
would result in an empty position in the corresponding std::vector,
and verifyFI() will fail.
So, one should not do that except when it's part of replace() or such.

* There is support for removing the entire ForallIntents clause
from the AST: use removeFI().

* Added AST_CALL_STDVEC as a counterpart to AST_CALL_LIST to traverse
a std::vector.

* for_vector finishes iteration once it hits a null vector element.
Since my riSpecs is expected to contain nulls, I added
for_vector_allowing_0s and a wrapper around it, for_riSpecs_vector.

* Cloned markOuterVarsWithIntents() for forall intents to use
ForallIntents, which the task-intent version still does not use.

* WHILE THERE: removed INT_ASSERT(!byrefVars) from codegen functions.
This is done in verify() instead.

* WHILE THERE: moved the "default" clause out of the switch in qualStr()
and in tiMarkForIntent() to enable compiler-compile-time check for
"did we cover all cases".

EXPERIENCE
==========

ForallIntents is a lot like adding a new AST node.
What issues did I encounter in the process?

* I chose ForallIntents not to be an actual AST node. The key argument
is that eventually ForallIntents's fields will be directly under
an upcoming ForallStmt node. The current structure resembles that.
Having ForallIntents as a node wouldn't, as its fields would list
the ForallIntents node as their parentExpr. Also see the following...

* For each 'block', expressions immediately under block->forallIntents
have 'block' as their parentExpr. This caused trouble because that
made them be considered "statement-level expressions".
E.g. normalize() would try to add _statementLevelSymbol around those.
To address that, I adjusted normalize() and getStmtExpr()/isStmtExpr().

* ForallIntents needs to support AST operations of its BlockStmt parent,
e.g.:
   - copy
   - visitor
   - traversal
   - adding to/removal from the tree; destruction
   - removal
       actually, only 2 things to support: parent BlockStmt->remove()
       and parent BlockStmt removing the ForallIntents
   - verify

It was a bit odd to implement these for a non-AST node.
Now that it's done, however, we do not even need to think about it.

* The parser needs ForallIntents. I did not dare to #include foralls.h
into chapel.ypp. So I added a couple of ForallIntents funs to symbol.h
so the parser sees them.

* I needed to augment all AstVisitor subclasses to handle ForallIntents.
This makes sense, as this is code that is affected by forall intents.

* I chose to use std::vector instead of AList for lists of forall-intent
variables and reduce specs - based on my vague concern aboutAList::parent.
In retrospect this probably should not be a concern.

Because of std::vector, I had to add AST_CALL_STDVEC
and use the more verbose std::vector::iterator-based iteration
in some cases.

Also in retrospect/possibly TODO: could combine (variable, tag, reduce spec)
into a struct so that there is only a single vector.

Here are my current-thinking pros for each option:

std::vector :
- "modern" data structure
- convenient to iterate over the three lists in parallel
- can't use AList if (variable, tag, reduce spec) are in a single struct
- explicit support for the absence of a reduce spec (i.e. non-reduce intents)
  - just store null

AList :
- It is fast to check whether some Expr is on the list.